add new test, still working on custom validator
